refactor: move DRM reducer to its own file
